# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InvestPro Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (ban, suspend, balance)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/deposits:
    get:
      operationId: listAdminDeposits
      tags: [admin]
      summary: List all deposits
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDeposit"
  /admin/deposits/{id}/approve:
    post:
      operationId: approveDeposit
      tags: [admin]
      summary: Approve a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/deposits/{id}/reject:
    post:
      operationId: rejectDeposit
      tags: [admin]
      summary: Reject a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/withdrawals:
    get:
      operationId: listAdminWithdrawals
      tags: [admin]
      summary: List all withdrawals
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminWithdrawal"
  /admin/withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/kyc:
    get:
      operationId: listAdminKyc
      tags: [admin]
      summary: List all KYC submissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: KYC list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminKycItem"
  /admin/kyc/{id}/approve:
    post:
      operationId: approveKyc
      tags: [admin]
      summary: Approve KYC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
  /admin/kyc/{id}/reject:
    post:
      operationId: rejectKyc
      tags: [admin]
      summary: Reject KYC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
  /admin/plans:
    post:
      operationId: createPlan
      tags: [admin]
      summary: Create an investment plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    patch:
      operationId: updatePlan
      tags: [admin]
      summary: Update a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/tickets:
    get:
      operationId: listAdminTickets
      tags: [admin]
      summary: List all support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /admin/tickets/{id}/reply:
    post:
      operationId: adminReplyTicket
      tags: [admin]
      summary: Admin reply to ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "201":
          description: Reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
  /admin/tickets/{id}/close:
    post:
      operationId: closeTicket
      tags: [admin]
      summary: Close a support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
components:
  schemas:
...
    User:
      type: object
      required: [id, firstName, lastName, email, phone, country, role, status, kycLevel, vipLevel, referralCode, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, banned]
        kycLevel:
          type: integer
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        vipLevel:
          type: integer
        referralCode:
          type: string
        totalInvested:
          type: number
        createdAt:
          type: string
...
    Plan:
      type: object
      required: [id, name, minDeposit, maxDeposit, dailyRate, durationDays, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        minDeposit:
          type: number
        maxDeposit:
          type: ["number", "null"]
        dailyRate:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
        features:
          type: array
          items:
            type: string
    PlanInput:
      type: object
      required: [name, minDeposit, dailyRate, durationDays]
      properties:
        name:
          type: string
        description:
          type: string
        minDeposit:
          type: number
        maxDeposit:
          type: ["number", "null"]
        dailyRate:
          type: number
        durationDays:
          type: integer
        features:
          type: array
          items:
            type: string
...
    Deposit:
      type: object
      required: [id, amount, status, walletAddress, txHash, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        walletAddress:
          type: string
        txHash:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Withdrawal:
      type: object
      required: [id, amount, fee, netAmount, status, walletAddress, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        walletAddress:
          type: string
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
...
    KycStatus:
      type: object
      required: [level, status]
      properties:
        level:
          type: integer
        status:
          type: string
          enum: [none, pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        documents:
          type: array
          items:
            $ref: "#/components/schemas/KycDocument"
...
    TicketReplyInput:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Admin
...
    AdminUserList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUserDetail:
      type: object
      required: [user, wallet, investments, deposits, withdrawals]
      properties:
        user:
          $ref: "#/components/schemas/User"
        wallet:
          $ref: "#/components/schemas/Wallet"
        investments:
          type: array
          items:
            $ref: "#/components/schemas/Investment"
        deposits:
          type: array
          items:
            $ref: "#/components/schemas/Deposit"
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
    AdminUserUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended, banned]
        mainBalance:
          type: number
        bonusBalance:
          type: number
        kycLevel:
          type: integer
    AdminDeposit:
      type: object
      required: [id, userId, userName, amount, status, txHash, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        status:
          type: string
        txHash:
          type: ["string", "null"]
        walletAddress:
          type: string
        createdAt:
          type: string
    AdminWithdrawal:
      type: object
      required: [id, userId, userName, amount, fee, netAmount, status, walletAddress, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        status:
          type: string
        walletAddress:
          type: string
        createdAt:
          type: string
    AdminKycItem:
      type: object
      required: [id, userId, userName, documentType, status, submittedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        documentType:
          type: string
        status:
          type: string
        submittedAt:
          type: string
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string